% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{compute_mse_loss}
\alias{compute_mse_loss}
\title{MSE loss for data (y_i - pool_mean_i)^2 for i in pool i.}
\usage{
compute_mse_loss(
  data,
  value,
  M,
  sigma,
  policy_list,
  reg = 1,
  normalize = 0,
  lattice_edges = NA,
  return_dict = TRUE
)
}
\arguments{
\item{data}{Data.table containing the treatment assignments}

\item{value}{Label of column containing value observed for each individual}

\item{M}{Data.table of policy means}

\item{sigma}{Partition matrix that gives pooling structure}

\item{policy_list}{List of policies, which if a list of vectors of all policies implied
by the data.}

\item{reg}{Regularization parameter that penalizes partitions with more pools}

\item{normalize}{Whether or not to normalize loss}

\item{lattice_edges}{Edges of pooling structure}

\item{return_dict}{Whether or not to return the dictionary object that gives the pool
means for each model in the RashomonSet. Defaults to True.}
}
\value{
MSE given sigma pooling structure and data.
}
\description{
MSE loss for data (y_i - pool_mean_i)^2 for i in pool i.
}
