% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{extract_pools}
\alias{extract_pools}
\title{Helper function to extract pools from a given sigma matrix}
\usage{
extract_pools(policy_list, sigma, lattice_edges = NA)
}
\arguments{
\item{policy_list}{A list of policies}

\item{sigma}{Row of the sigma matrix to cut at}

\item{lattice_edges}{List of edges between policies given a pooling structure.}
}
\value{
A collections:dict object of all the pools, where the key is integer
id of a policy and the value is the pool id.
}
\description{
Helper function to extract pools from a given sigma matrix
}
