% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{find_R}
\alias{find_R}
\title{Finds R, a list (or integer) of the number of levels in each arm.}
\usage{
find_R(sigma)
}
\arguments{
\item{sigma}{Partition matrix for a given pooling structure}
}
\value{
A list (or integer) of the number of levels in each arm.
}
\description{
Finds R, a list (or integer) of the number of levels in each arm.
}
