% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rashomon_set.R
\name{find_rashomon_profile}
\alias{find_rashomon_profile}
\title{Finds the RashomonSet for a given profile.}
\usage{
find_rashomon_profile(
  data,
  value,
  arm_cols = c(),
  M,
  R,
  H,
  reg = 1,
  profile,
  policies = c(),
  policy_means = c(),
  normalize = 0,
  theta,
  filtered = FALSE,
  inactive = 0
)
}
\arguments{
\item{data}{A dataframe containing the column whose name you supply in value.}

\item{value}{The column name of the y values in data}

\item{arm_cols}{A character vector containing the names of the arm columns}

\item{M}{The number of arms}

\item{R}{An integer (or vector) denoting the number of levels in each arm. If
this value is an integer, it assumes each arm has the same number of levels.}

\item{H}{The maximum number of pools allowed to be specified by a partition matrix}

\item{reg}{The value of the regularization penalty when computing the loss for
a pooling structure. Defaults to 1, so that each additional pool in a model adds 1 to the loss.}

\item{profile}{Binary vector that denotes which arms are active.}

\item{policies}{A list of policies present in the data. This should be the output
of create_policies_from_data.}

\item{policy_means}{A dataframe that gives the mean value of value for each policy.}

\item{normalize}{Normalization factor. When this is 0, no normalization is applied. Should be set
to the size of the total dataset, so that each loss represents contribution to the overall
loss of the model across all profiles. Defaults to zero.}

\item{theta}{Threshold value to be present in the RashomonSet. Set to Inf if you want all poolings.}

\item{filtered}{Whether or not data is already filtered to be only the observations in the given
profile. Defaults to True.}

\item{inactive}{The level that denotes an inactive arm in data. Defaults to zero.}
}
\value{
A RashomonSet Object that gives all of the partition matrices in the
Rashomon Set for a given theta.
}
\description{
Finds the RashomonSet for a given profile.
}
