% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{policy_means}
\alias{policy_means}
\title{Finds policy labels given data with each observation labeled with its policy
TODO: Change all function calls from other packages to be explicit, and
change if necessary for speedups.}
\usage{
policy_means(data, value)
}
\arguments{
\item{data}{Data.table containing the treatment assignments}

\item{value}{Name of value column observed for each observation, supplied by the user. Must be character.}
}
\value{
A Data.table of policy means, where each row is the mean for a given policy.
}
\description{
Finds policy labels given data with each observation labeled with its policy
TODO: Change all function calls from other packages to be explicit, and
change if necessary for speedups.
}
