% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{prep_data}
\alias{prep_data}
\title{Prep data for use in raggregate_profiles() or find_rashomon_profile()}
\usage{
prep_data(data, arm_cols, value, R, drop_unobserved_combinations = FALSE)
}
\arguments{
\item{data}{A dataframe containing the column whose name you supply in value}

\item{arm_cols}{A character vector containing the names of the arm columns}

\item{value}{The column name of the y values in data}

\item{R}{An integer (or vector) denoting the number of levels in each arm. If
this value is an integer, it assumes each arm has the same number of levels.}

\item{drop_unobserved_combinations}{Whether or not to drop factor combinations
not observed in the data. If false, will return a dataframe with all factor
combinations labelled, with NAs in value column for factor combinations implicitly missing.}
}
\description{
Preps data for use in core functions.
}
