% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RashomonSet.R
\name{sort_rashomon}
\alias{sort_rashomon}
\title{Sorts this rashomon set according to the model with the smallest loss.}
\usage{
sort_rashomon(obj)
}
\arguments{
\item{obj}{A RashomonSet object}
}
\description{
Sorts this rashomon set according to the model with the smallest loss.
}
